% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_algo_light}
\alias{palette_algo_light}
\alias{palette_algo_dark}
\title{Algoritma color palette}
\usage{
palette_algo_light(color = NULL, colour = color)

palette_algo_dark(color = NULL, colour = color)
}
\arguments{
\item{color, colour}{Specific color that need to be returned. See details.}
}
\description{
Algoritma color palette
}
\details{
Will be updated soon.
}
\examples{

# import libs
library(algothemes)

# get the full light palette
palette_algo_light()

# get the full dark palette
palette_algo_dark()

}
