% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color.R
\name{scale_color_algo}
\alias{scale_color_algo}
\alias{scale_colour_algo}
\title{Algoritma color scale}
\usage{
scale_color_algo(discrete = TRUE, mode = "light", ...)

scale_colour_algo(discrete = TRUE, ...)
}
\arguments{
\item{discrete}{A \code{logical} to specify whether to scale the colour in
discrete or continuous scale. Default to \code{TRUE}.}

\item{mode}{The theme's mode; either \code{"light"} or \code{"dark"} mode.}

\item{...}{Other arguments passed on to \code{discrete_scale()}
or \code{scale_color_gradientn()}.}
}
\description{
Algoritma color scale
}
\details{
Will be updated soon.
}
\examples{

# import libs
library(ggplot2)
library(algothemes)

# an example visualization
p <- ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point(aes(colour = Species))

p

# with theme_algo and scale_colour_algo
p +
  theme_algo() +
  scale_colour_algo()

}
