% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_algo}
\alias{theme_algo}
\title{Algoritma themes}
\usage{
theme_algo(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22,
  mode = "light")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{mode}{The theme's mode; either \code{"light"} or \code{"dark"} mode.}
}
\description{
Algoritma themes
}
\details{
Will be updated soon.
}
\examples{

# import libs
library(ggplot2)
library(algothemes)

# an example visualization
p <- ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point(aes(colour = Species))

p

# with theme_algo
p +
  theme_algo()

}
